Deduplicate clock control CPU software
