+++ firmware-support/bittide-cpus/src/lib.rs
@@ -5,3 +5,4 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod boot;
+pub mod clock_control;
